+++ xunit.runner.worker/DiscoverUtil.cs
@@ -1,49 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.IO;
 using Xunit;
 using Xunit.Abstractions;
 using xunit.runner.data;
-using System.Threading;
 
 namespace xunit.runner.worker
 {
     internal sealed class DiscoverUtil
     {
-        private sealed class Impl : LongLivedMarshalByRefObject, IMessageSink, IDisposable
+        private sealed class Impl : TestDiscoverySink
         {
-            private readonly ITestFrameworkDiscoverer _discoverer;
             private readonly ClientWriter _writer;
 
-            public ManualResetEvent Finished { get; private set; }
-
-            internal Impl(ITestFrameworkDiscoverer discoverer, ClientWriter writer)
+            internal Impl(ClientWriter writer)
             {
-                Finished = new ManualResetEvent(false);
-                _discoverer = discoverer;
                 _writer = writer;
             }
 
-            public bool OnMessage(IMessageSinkMessage message)
-            {
-                var discoveryMessage = message as ITestCaseDiscoveryMessage;
-                if (discoveryMessage != null)
-                {
-                    OnDiscoveryMessage(discoveryMessage);
-                }
+            protected override bool ShouldContinue => _writer.IsConnected;
 
-                if (message is IDiscoveryCompleteMessage)
-                {
-                    Finished.Set();
-                }
-
-                return _writer.IsConnected;
-            }
-
-            private bool OnDiscoveryMessage(ITestCaseDiscoveryMessage testCaseDiscovered)
+            protected override void OnTestDiscovered(ITestCaseDiscoveryMessage testCaseDiscovered)
             {
                 var testCase = testCaseDiscovered.TestCase;
                 var testCaseData = new TestCaseData(
@@ -53,13 +28,6 @@ namespace xunit.runner.worker
 
                 _writer.Write(TestDataKind.Value);
                 _writer.Write(testCaseData);
-
-                return _writer.IsConnected;
-            }
-
-            public void Dispose()
-            {
-                ((IDisposable)Finished).Dispose();
             }
         }
 
@@ -72,7 +40,7 @@ namespace xunit.runner.worker
                 diagnosticMessageSink: new MessageVisitor(),
                 shadowCopy: false))
             using (var writer = new ClientWriter(stream))
-            using (var impl = new Impl(xunit, writer))
+            using (var impl = new Impl(writer))
             {
                 xunit.Find(includeSourceInformation: false, messageSink: impl, discoveryOptions: TestFrameworkOptions.ForDiscovery());
                 impl.Finished.WaitOne();
