+++ xunit.runner.worker/xunit.runner.worker.csproj
@@ -58,6 +58,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RunUtil.cs" />
+    <Compile Include="TestDiscoverySink.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
